feat: unsub on channel info modal
